feat(loom): support workspace loom builds
Add the loom thread/runtime shims, stack sizing, deterministic test
hooks, and cfg gates needed for workspace crates to compile under
`features=loom`.

Move the dataplane runtime verbatim from `main.rs` into `runtime.rs`
so the binary can be stubbed out for loom builds. The relocation is
purely mechanical -- the moved code is unchanged apart from fixing the
driver import path (`drivers::` -> `crate::drivers::`) and making
`main` `pub` so `main.rs` can call `runtime::main()`. No logic was
edited.
