+++ concurrency/Cargo.toml
@@ -9,7 +9,6 @@ version.workspace = true
 default = ["parking_lot"]
 loom = ["dep:loom", "concurrency-macros/loom"]
 parking_lot = ["dep:parking_lot"]
-#
 shuttle = ["dep:shuttle", "concurrency-macros/shuttle"]
 shuttle_dfs = ["shuttle"]
 silence_clippy = ["concurrency-macros/silence_clippy"] # Do not manually enable this feature, let --all-features do it for you
