+++ concurrency/src/stress.rs
@@ -8,13 +8,14 @@
 //!
 //! * default backend -- direct call, no scheduling exploration.
 //! * `loom` feature -- `loom::model`.
-//!
-//!
+//! * `shuttle` feature -- a [`shuttle::PortfolioRunner`] that runs
+//!   `RandomScheduler` and `PctScheduler` in parallel;
+//! * `shuttle_dfs` (additive on top of `shuttle`) -- also adds
+//!   `DfsScheduler` to the portfolio.
 
 /// Run `body` under the currently selected concurrency backend.
-///
-/// See the module docs for the per-backend dispatch table.
 #[allow(unused_variables)] // `body` may be unused in arms that don't take a closure.
+#[allow(clippy::expect_used)] // backend spawn / join: panic-on-failure is the right semantic in a test harness.
 pub fn stress<F>(body: F)
 where
     F: Fn() + Send + Sync + 'static,
@@ -24,8 +25,22 @@ where
     #[cfg(all(not(feature = "loom"), feature = "shuttle"))]
     const SCHEDULES: usize = 3;
 
+    #[cfg(feature = "loom")]
+    const LOOM_STACK_SIZE: usize = 4 * 1024 * 1024;
+
     cfg_select! {
-        feature = "loom" => { loom::model(body); },
+        feature = "loom" => {
+            let body = std::sync::Arc::new(body);
+            loom::model(move || {
+                let body = body.clone();
+                loom::thread::Builder::new()
+                    .stack_size(LOOM_STACK_SIZE)
+                    .spawn(move || body())
+                    .expect("loom thread spawn")
+                    .join()
+                    .expect("loom body panicked");
+            });
+        },
         feature = "shuttle" => {
             use shuttle::PortfolioRunner;
             use shuttle::scheduler::{PctScheduler, RandomScheduler};
@@ -44,11 +59,6 @@ where
             portfolio.run(body);
         },
         not(any(feature = "loom", feature = "shuttle")) => { body(); },
-        _ => compile_error!(
-            "stress: a model-checker feature is enabled but no dispatch \
-             arm matched. Either an explicit arm above is missing, or \
-             the `not(any(...))` default needs widening to cover the \
-             new feature.",
-        ),
+        _ => compile_error!("stress: no dispatch arm matched the active feature set"),
     }
 }
