+++ concurrency/src/sync/mod.rs
@@ -12,14 +12,13 @@
 //!   a local `Arc<T>` / `Weak<T>` shim (loom 0.7 ships `Arc` but no
 //!   `Weak`).
 //! * `shuttle` feature (plus the additive `shuttle_dfs` opt-in):
+//!   poison-as-panic wrapper around `shuttle::sync`.
 //! * `parking_lot` feature (default): zero-cost re-export of
 //!   `parking_lot`'s naked-guard locks; the production hot path.
 //!   Skipped when `_strict_provenance` is on, even if `parking_lot`
 //!   is also on, because `parking_lot_core::word_lock` uses
 //!   integer-to-pointer casts that miri's strict-provenance mode
-//!   rejects; the CI miri job exercises the fallback slot under
-//!   strict provenance, and that needs the sync surface to come
-//!   from `std::sync`.
+//!   rejects.
 //! * Otherwise: `std_backend` -- a thin poison-as-panic wrapper around
 //!   `std::sync`. Lets `--no-default-features` and
 //!   `_strict_provenance` builds compile without depending on
@@ -28,10 +27,8 @@
 //! # Portability footguns the facade *does not* paper over
 //!
 //! The wrapped backends are observationally compatible with the
-//! production `parking_lot` surface for the things call sites
-//! actually use, but a few API details diverge in ways that matter
-//! to anyone writing a static, a model-checked test, or code that
-//! relies on `parking_lot`-specific schedules:
+//! production `parking_lot` surface for current call sites, but a few
+//! API details still diverge:
 //!
 //! * **`Mutex::new` / `RwLock::new` are not `const fn` under
 //!   `loom`/`shuttle*`.** loom's `Mutex::new` is plain `fn` because
@@ -40,26 +37,20 @@
 //!   denominator. So `static M: Mutex<T> = Mutex::new(...)` compiles
 //!   under the default and `parking_lot` backends and fails to
 //!   typecheck under the model-checker backends. Workaround for
-//!   tests that need a static: wrap the static in `OnceLock`, or
-//!   construct the `Mutex` inside the test body.
+//!   tests that need a static: wrap the static in `OnceLock`.
 //!
 //! * **`OnceLock` under `loom`/`shuttle*` is re-exported from
 //!   `std::sync` unchanged.** It is sound for laziness, but it uses
 //!   uninstrumented atomics inside, so the model checker does *not*
 //!   explore the orderings around `OnceLock::get_or_init`. Tests
-//!   whose correctness depends on the publication ordering of a
-//!   once-initialised cell need to model that ordering explicitly
-//!   (e.g. an `Arc<T>` + an explicit `Acquire` load on the
-//!   subscriber, both of which loom *does* model).
+//!   whose correctness depends on that publication ordering need to
+//!   model it explicitly.
 //!
 //! * **`RwLock::upgradable_read` under `loom`/`shuttle*` takes an
 //!   exclusive write lock.** Sound -- no schedule that `parking_lot`
 //!   would allow is forbidden -- but lossy: the model checker never
 //!   explores the many-readers-plus-one-upgradable schedule that
-//!   `parking_lot` permits. Code whose correctness hinges on that
-//!   specific interleaving needs an explicit `read()` then `write()`
-//!   pair (which loom *can* model), or a richer state machine in
-//!   the facade.
+//!   `parking_lot` permits.
 
 // loom takes priority so the model checker can drive its own internal
 // state (used for tests that opt loom in explicitly).
@@ -101,10 +92,7 @@ mod std_backend;
 ))]
 pub use std_backend::*;
 
-// Match the silence_clippy escape hatch in lib.rs: when both loom and
-// shuttle are pulled in (under `--all-features`), route sync through
-// `std` purely to keep clippy happy. The binary is never executed in
-// that configuration.
+// Type-check only; never executed under `--all-features`.
 #[cfg(all(feature = "shuttle", feature = "loom", feature = "silence_clippy"))]
 mod std_backend;
 #[cfg(all(feature = "shuttle", feature = "loom", feature = "silence_clippy"))]
