+++ concurrency/src/thread/mod.rs
@@ -3,11 +3,6 @@
 
 //! Backend-routed threading primitives.
 //!
-//! Re-exports the active backend's `thread` module wholesale (`spawn`,
-//! `current`, `sleep`, `yield_now`, `JoinHandle`, `Thread`, `ThreadId`,
-//! `Builder`, ...) so call sites use one path regardless of whether
-//! they're building against `std`, `loom`, or `shuttle`.
-//!
 //! ## `thread::scope`
 //!
 //! `std::thread::scope` (stable since 1.63) and `shuttle::thread::scope`
@@ -31,7 +26,35 @@ pub use std::thread::*;
 pub use shuttle::thread::*;
 
 #[cfg(all(feature = "loom", not(feature = "silence_clippy")))]
-pub use loom::thread::*;
+pub use loom::thread::{
+    AccessError, Builder, JoinHandle, LocalKey, Thread, ThreadId, current, panicking, park,
+    yield_now,
+};
+
+/// Spawn a loom thread
+///
+/// # Panics
+/// Panics if loom's scheduler refuses the spawn (the underlying
+/// `Builder::spawn` returns `io::Result`; for loom this is effectively
+/// infallible, so the `expect` is a stand-in for `unwrap_unchecked`).
+#[cfg(all(feature = "loom", not(feature = "silence_clippy")))]
+#[allow(clippy::expect_used)]
+pub fn spawn<F, T>(f: F) -> JoinHandle<T>
+where
+    F: FnOnce() -> T + Send + 'static,
+    T: Send + 'static,
+{
+    Builder::new()
+        .stack_size(4 * 1024 * 1024)
+        .spawn(f)
+        .expect("loom thread spawn")
+}
+
+/// Backend-portable `sleep` shim under loom.
+#[cfg(all(feature = "loom", not(feature = "silence_clippy")))]
+pub fn sleep(_: core::time::Duration) {
+    loom::thread::yield_now();
+}
 
 #[cfg(all(feature = "loom", not(feature = "silence_clippy")))]
 mod loom_scope;
@@ -39,13 +62,12 @@ mod loom_scope;
 #[cfg(all(feature = "loom", not(feature = "silence_clippy")))]
 pub use loom_scope::{Scope, ScopedJoinHandle, scope};
 
-// Match the silence_clippy escape hatch in `crate::sync`: under
-// `--all-features` both loom and shuttle are enabled at once, which
-// can't pick a single backend. Route to `std::thread` so the binary
-// type-checks; it is never executed in that configuration.
+// Type-check only; never executed under `--all-features`.
 #[cfg(all(feature = "shuttle", feature = "loom", feature = "silence_clippy"))]
 pub use std::thread::*;
 
+// `Builder::spawn_scoped` is only inherent on `std::thread::Builder`;
+// loom/shuttle provide `Scope::spawn` instead. This trait closes the gap.
 pub trait BuilderExt {
     /// Spawn a thread within `scope`.
     ///
