+++ nat/src/stateful/apalloc/port_alloc.rs
@@ -161,6 +161,11 @@ impl<I: NatIpWithBitmap> PortAllocator<I> {
         }
     }
 
+    #[concurrency_mode(loom)]
+    fn shuffle_slice<T>(_slice: &mut [T]) {
+        // Deterministic replay matters more than allocation-order randomization.
+    }
+
     // Iterate over the slice of all blocks, but starting from a given offset (and looping at the
     // end), returning the block and its index from the initial slice.
     //
